Jammoney -> jammoney
