--- src/main/java/com/example/Jammoney/StockApp/kis/dto/KospiDto.java
@@ -1,13 +1,0 @@
-package com.example.Jammoney.StockApp.kis.dto;
-
-import lombok.Data;
-import java.time.LocalDate;
-
-@Data
-public class KospiDto {
-
-    private LocalDate date;
-    private int close;
-    private int prdyVrss;
-    private double prdyCrt;
-}
