--- src/main/java/com/example/Jammoney/StockApp/kis/entity/KospiRepository.java
@@ -1,9 +1,0 @@
-package com.example.Jammoney.StockApp.kis.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface KospiRepository extends JpaRepository<Kospi, Long> {
-    List<Kospi> findAllByOrderByDateAsc();
-}
