--- src/main/java/com/example/Jammoney/StockApp/kis/repository/KisTokenRepository.java
@@ -1,7 +1,0 @@
-package com.example.Jammoney.StockApp.kis.repository;
-import com.example.Jammoney.StockApp.kis.entity.KisToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface KisTokenRepository extends JpaRepository<KisToken, Long> {
-}
-
