--- src/main/java/com/example/Jammoney/StockApp/stock/dto/InterestingStockResponseDto.java
@@ -1,11 +1,0 @@
-package com.example.Jammoney.StockApp.stock.dto;
-
-public class InterestingStockResponseDto {
-    private long companyId;
-    private String companyName;
-    private String currentPrice;
-    private String priceChange;
-    private String priceChangeRate;
-    private String volume;
-    private String tradeValue;
-}
