--- src/main/java/com/example/Jammoney/StockApp/stock/dto/OrderResponseDto.java
@@ -1,18 +1,0 @@
-package com.example.Jammoney.StockApp.stock.dto;
-
-import com.example.Jammoney.StockApp.stock.entity.Enums.OrderStatus;
-import com.example.Jammoney.StockApp.stock.entity.Enums.OrderType;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class OrderResponseDto {
-    long stockOrderId;
-    int stockCount;
-    long userId;
-    long companyId;
-    OrderStatus orderStatus;
-    OrderType orderType;
-    long price;
-}
