--- src/main/java/com/example/Jammoney/StockApp/stock/entity/Enums/OrderStatus.java
@@ -1,6 +1,0 @@
-package com.example.Jammoney.StockApp.stock.entity.Enums;
-
-public enum OrderStatus {
-    COMPLETED,
-    WAITING
-}
