--- src/main/java/com/example/Jammoney/StockApp/stock/entity/Enums/OrderType.java
@@ -1,6 +1,0 @@
-package com.example.Jammoney.StockApp.stock.entity.Enums;
-
-public enum OrderType {
-    BUY,
-    SELL
-}
