--- src/main/java/com/example/Jammoney/StockApp/stock/entity/Order.java
@@ -1,41 +1,0 @@
-package com.example.Jammoney.StockApp.stock.entity;
-
-import com.example.Jammoney.StockApp.stock.entity.Enums.OrderStatus;
-import com.example.Jammoney.StockApp.stock.entity.Enums.OrderType;
-import com.example.Jammoney.User.User;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class Order{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long stockOrderId;
-
-    @Column
-    private int stockCount;
-
-    @ManyToOne()
-    @JoinColumn(name = "MEMBER_ID")
-    @OnDelete(action = OnDeleteAction.CASCADE)
-    private User user;
-    @ManyToOne()
-    @JoinColumn(name = "COMPANY_ID")
-    private Company company;
-
-    @Enumerated(EnumType.STRING) // 항상 명시하자!
-    private OrderStatus orderStatus;
-
-    @Enumerated(EnumType.STRING) // 항상 명시하자!
-    private OrderType orderType;
-
-    private long price;
-
-}
