+++ src/main/java/com/example/jammoney/StockApp/kis/entity/KisToken.java
@@ -1,0 +1,23 @@
+package com.example.jammoney.StockApp.kis.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class KisToken {
+
+    @Id
+    private Long id = 1L; // 단일 행만 저장
+
+    @Column(nullable = false)
+    private String token;
+
+    @Column(nullable = false)
+    private LocalDateTime expiredAt;
+}
