+++ src/main/java/com/example/jammoney/StockApp/kis/entity/KospiRepository.java
@@ -1,0 +1,9 @@
+package com.example.jammoney.StockApp.kis.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface KospiRepository extends JpaRepository<Kospi, Long> {
+    List<Kospi> findAllByOrderByDateAsc();
+}
