+++ src/main/java/com/example/jammoney/StockApp/kis/repository/KisTokenRepository.java
@@ -1,0 +1,7 @@
+package com.example.jammoney.StockApp.kis.repository;
+import com.example.jammoney.StockApp.kis.entity.KisToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface KisTokenRepository extends JpaRepository<KisToken, Long> {
+}
+
