+++ src/main/java/com/example/jammoney/StockApp/stock/dto/InterestingStockResponseDto.java
@@ -1,0 +1,11 @@
+package com.example.jammoney.StockApp.stock.dto;
+
+public class InterestingStockResponseDto {
+    private long companyId;
+    private String companyName;
+    private String currentPrice;
+    private String priceChange;
+    private String priceChangeRate;
+    private String volume;
+    private String tradeValue;
+}
