+++ src/main/java/com/example/jammoney/StockApp/stock/entity/Enums/OrderStatus.java
@@ -1,0 +1,6 @@
+package com.example.jammoney.StockApp.stock.entity.Enums;
+
+public enum OrderStatus {
+    COMPLETED,
+    WAITING
+}
