+++ src/main/java/com/example/jammoney/StockApp/stock/entity/InterestingStock.java
@@ -1,0 +1,27 @@
+package com.example.jammoney.StockApp.stock.entity;
+import com.example.jammoney.User.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class InterestingStock {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long interestingStockId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "MEMBER_ID")
+    @OnDelete(action = OnDeleteAction.CASCADE)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "COMPANY_ID")
+    private Company company;
+}
